% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices-list.R
\name{devices_list}
\alias{devices_list}
\title{Devices list API methods}
\usage{
devices_list(
  serial = NULL,
  tags = NULL,
  token = getOption("ARDUINO_API_TOKEN")
)
}
\arguments{
\item{serial}{serial number of the device you  may want to filter from the list (not device_id)}

\item{tags}{tags you  may want to filter from the list}

\item{token}{A valid token created with \code{create_auth_token}
(either explicitely assigned or retrieved via default \code{getOption('ARDUINO_API_TOKEN')})}
}
\value{
A tibble showing extensive information about devices (and related things) associated to the user
}
\description{
List devices associated to the user
}
\details{
Official documentation:
\itemize{
\item \href{https://www.arduino.cc/reference/en/iot/api/#api-DevicesV2-devicesV2List}{devicesV2List}
}
}
\examples{
\dontrun{
Sys.setenv(ARDUINO_API_CLIENT_ID = 'INSERT CLIENT_ID HERE')
Sys.setenv(ARDUINO_API_CLIENT_SECRET = 'INSERT CLIENT_SECRET HERE')

create_auth_token()

devices_list()
}
}
