% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices-properties.R
\name{devices_properties}
\alias{devices_properties}
\alias{devices_properties_list}
\title{Properties (of devices) API methods}
\usage{
devices_properties_list(
  device_id,
  show_deleted = FALSE,
  token = getOption("ARDUINO_API_TOKEN")
)
}
\arguments{
\item{device_id}{The id of the device}

\item{show_deleted}{If \code{TRUE}, shows the soft deleted properties. Default to \code{FALSE}}

\item{token}{A valid token created with \code{create_auth_token}
(either explicitly assigned or retrieved via default \code{getOption('ARDUINO_API_TOKEN')})}
}
\value{
A tibble showing the information about properties for given device.
}
\description{
List properties associated to a given device
}
\details{
Official documentation: \href{https://www.arduino.cc/reference/en/iot/api/#api-DevicesV2-devicesV2GetProperties}{devicesV2GetProperties}
}
\examples{
\dontrun{
library(dplyr)

Sys.setenv(ARDUINO_API_CLIENT_ID = 'INSERT CLIENT_ID HERE')
Sys.setenv(ARDUINO_API_CLIENT_SECRET = 'INSERT CLIENT_SECRET HERE')
create_auth_token()

device_id = "fa7ee291-8dc8-4713-92c7-9027969e4aa1"

### check properties list ###
devices_properties_list(device_id = device_id)
}
}
