% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devices-tags.R
\name{devices_tags}
\alias{devices_tags}
\alias{devices_tags_upsert}
\alias{devices_tags_list}
\alias{devices_tags_delete}
\title{Tags (of devices) API methods}
\usage{
devices_tags_upsert(
  device_id,
  key,
  value,
  token = getOption("ARDUINO_API_TOKEN")
)

devices_tags_list(device_id, token = getOption("ARDUINO_API_TOKEN"))

devices_tags_delete(device_id, key, token = getOption("ARDUINO_API_TOKEN"))
}
\arguments{
\item{device_id}{The id of the device}

\item{key}{The key of the tag (no spaces allowed)}

\item{value}{The value of the tag (no spaces allowed)}

\item{token}{A valid token created with \code{create_auth_token}
(either explicitely assigned or retrieved via default getOption('ARDUINO_API_TOKEN'))}
}
\value{
A tibble showing the keys and values for given device
}
\description{
Upsert (create/update), List and Delete tags associated to a given device
}
\details{
Official documentation:
\itemize{
\item \href{https://www.arduino.cc/reference/en/iot/api/#api-DevicesV2Tags-devicesV2TagsUpsert}{devicesV2TagsUpsert}
\item \href{https://www.arduino.cc/reference/en/iot/api/#api-DevicesV2Tags-devicesV2TagsList}{devicesV2TagsList}
\item \href{https://www.arduino.cc/reference/en/iot/api/#api-DevicesV2Tags-devicesV2TagsDelete}{devicesV2TagsDelete}
}
}
\examples{
Sys.setenv(ARDUINO_API_CLIENT_ID = 'V8CpJ82mOtpsBgnGqeVGvRpSw9SOXcNo')
Sys.setenv(ARDUINO_API_CLIENT_SECRET = 'OSo32AvUyJCe15UO90kXxrDQtsf1DpsFJ5CYI3xT9TaHe1SvIu8BwBteGl0pAdL5')
create_auth_token()

device_id = "fa7ee291-8dc8-4713-92c7-9027969e4aa1"
### create/modify tag ###
devices_tags_upsert(device_id = device_id, key = "1", value = "test")

### check tags list ###
devices_tags_list(device_id = device_id)

### delete tag ###
devices_tags_delete(device_id = device_id, key = "1")

}
