% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rduinoiot-package.R
\docType{package}
\name{Rduinoiot-package}
\alias{Rduinoiot}
\alias{Rduinoiot-package}
\title{Rduinoiot: Arduino iot-api R client}
\description{
This package provides an R client for the Arduino Iot Cloud API and some helpful functions that utilize the client.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FlavioLeccese92/Rduinoiot}
  \item Report bugs at \url{https://github.com/FlavioLeccese92/Rduinoiot/issues}
}

}
\keyword{internal}
