% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/things-list.R
\name{things_list}
\alias{things_list}
\title{Things list API methods}
\usage{
things_list(
  device_id = NULL,
  thing_id = NULL,
  show_deleted = FALSE,
  show_properties = FALSE,
  tags = NULL,
  token = getOption("ARDUINO_API_TOKEN")
)
}
\arguments{
\item{device_id}{device_id of the device you  may want to filter from the list}

\item{thing_id}{thing_id of the thing you may want to filter from the list}

\item{show_deleted}{(logical) If \code{TRUE}, shows the soft deleted things. Default to \code{FALSE}}

\item{show_properties}{(logical) If \code{TRUE}, returns things with their properties, and last values. Default to \code{FALSE}}

\item{tags}{tags you  may want to filter from the list}

\item{token}{A valid token created with \code{create_auth_token}
(either explicitely assigned or retrieved via default \code{getOption('ARDUINO_API_TOKEN')})}
}
\value{
A tibble showing information about things associated to the user
}
\description{
List things associated to the user
}
\details{
Official documentation:
\itemize{
\item \href{https://www.arduino.cc/reference/en/iot/api/#api-ThingsV2-thingsV2List}{thingsV2List}
}
}
\examples{
\dontrun{
Sys.setenv(ARDUINO_API_CLIENT_ID = 'INSERT CLIENT_ID HERE')
Sys.setenv(ARDUINO_API_CLIENT_SECRET = 'INSERT CLIENT_SECRET HERE')

create_auth_token()

t_list = things_list()
}
}
