% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-auth-token.R
\name{create_auth_token}
\alias{create_auth_token}
\title{Create Auth Token for Arduino IoT Cloud API}
\usage{
create_auth_token(
  client_id = Sys.getenv("ARDUINO_API_CLIENT_ID"),
  client_secret = Sys.getenv("ARDUINO_API_CLIENT_SECRET"),
  ...
)
}
\arguments{
\item{client_id}{Your client id (default is the environmental variable \code{ARDUINO_API_CLIENT_ID})}

\item{client_secret}{Your client secret (default is the environmental variable \code{ARDUINO_API_CLIENT_SECRET})}

\item{...}{Additional parameters needed for the body of the \code{POST} request:
\itemize{
\item \code{token_url} (default: \url{https://api2.arduino.cc/iot/v1/clients/token/})
\item \code{grant_type} (default: 'client_credentials')
\item \code{audience} (default: \url{https://api2.arduino.cc/iot/})
\item \code{content_type} (default: 'application/x-www-form-urlencoded')
}}
}
\value{
A token valid for Arduino IoT Cloud API (stored on .Rprofile) and retrievable by \code{getOption('ARDUINO_API_TOKEN')}
}
\description{
Obtain an authorization token using your personal client_id and client_secret
from Arduino site page \url{https://cloud.arduino.cc/home/api-keys/}

Official documentation: \href{https://www.arduino.cc/reference/en/iot/api/#api/}{API Summary}
}
\examples{
\dontrun{
# Sys.setenv(ARDUINO_API_CLIENT_ID = 'INSERT CLIENT_ID HERE')
# Sys.setenv(ARDUINO_API_CLIENT_SECRET = 'INSERT CLIENT_SECRET HERE')

create_auth_token()
}
}
